requirements_api.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
apache-airflow